finished expense test
